feat: Enhance performance logging across various components and update color palette handling in InitAgentLayer

+++ viewer_net/src/DanaSim.Viewer.Application/Dtos/InitAgentLayerDto.cs
@@ -15,5 +15,8 @@ public sealed class ColorPaletteEntryDto
     [JsonPropertyName("value")] public int Value { get; set; }
     [JsonPropertyName("label")] public string Label { get; set; } = string.Empty;
     [JsonPropertyName("hex")] public string Hex { get; set; } = string.Empty;
-    [JsonPropertyName("rgba")] public byte[] Rgba { get; set; } = [];
+    // The simulator sends `rgba` as a JSON array of integers (e.g. [158,128,80,255]),
+    // not a Base64 string — System.Text.Json's default byte[] converter only accepts
+    // the latter, so this must be int[] and converted to byte[] when consumed.
+    [JsonPropertyName("rgba")] public int[] Rgba { get; set; } = [];
 }
